Moved the downloadFiles function to an instance function
